yugiohbot/requirements.txt
facebook-sdk
requests